+++ packages/bootstrap-4/src/ObjectFieldTemplate/ObjectFieldTemplate.tsx
@@ -1,10 +1,15 @@
 import React from "react";
 
-import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
+import Container from "react-bootstrap/Container";
 
 import { ObjectFieldTemplateProps } from "@rjsf/core";
+import { utils } from '@rjsf/core';
+
+import AddButton from "../AddButton/AddButton";
+
+const { canExpand } = utils;
 
 const ObjectFieldTemplate = ({
   DescriptionField,
@@ -15,6 +20,11 @@ const ObjectFieldTemplate = ({
   required,
   uiSchema,
   idSchema,
+  schema,
+  formData,
+  onAddClick,
+  disabled,
+  readonly
 }: ObjectFieldTemplateProps) => {
   return (
     <>
@@ -40,7 +50,18 @@ const ObjectFieldTemplate = ({
             <Col xs={12}> {element.content}</Col>
           </Row>
         ))}
-      </Container>
+        {canExpand(schema, uiSchema, formData) ? (
+          <Row>
+            <Col xs={{ offset: 9, span: 3 }} className="py-4">
+              <AddButton
+                onClick={onAddClick(schema)}
+                disabled={disabled || readonly}
+                className="object-property-expand"
+              />
+            </Col>
+          </Row>
+        ) : null }
+      </Container> 
     </>
   );
 };