+++ packages/material-ui/src/FieldTemplate/WrapIfAdditional.tsx
@@ -1,8 +1,8 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { utils } from '@rjsf/core';
 import { JSONSchema7 } from 'json-schema';
 
-import MuiComponentContext from '../MuiComponentContext/MuiComponentContext';
+import { useMuiComponent } from '../MuiComponentContext';
 
 const { ADDITIONAL_PROPERTY_FLAG } = utils;
 
@@ -30,7 +30,7 @@ const WrapIfAdditional = ({
   required,
   schema,
 }: WrapIfAdditionalProps) => {
-  const { Grid, FormControl, IconButton, InputLabel, Input, RemoveIcon } = useContext(MuiComponentContext);
+  const { Grid, FormControl, IconButton, InputLabel, Input, RemoveIcon } = useMuiComponent();
   const keyLabel = `${label} Key`; // i18n ?
   const additional = schema.hasOwnProperty(ADDITIONAL_PROPERTY_FLAG);
   const btnStyle = {
@@ -44,8 +44,7 @@ const WrapIfAdditional = ({
     return <>{children}</>;
   }
 
-  const handleBlur = ({ target }: React.FocusEvent<HTMLInputElement>) =>
-    onKeyChange(target.value);
+  const handleBlur = ({ target }: React.FocusEvent<HTMLInputElement>) => onKeyChange(target.value);
 
   return (
     <Grid container={true} key={`${id}-key`} alignItems="center" spacing={2}>