+++ packages/material-ui/src/ObjectFieldTemplate/ObjectFieldTemplate.tsx
@@ -1,7 +1,7 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { ObjectFieldTemplateProps, utils } from '@rjsf/core';
 
-import MuiComponentContext from '../MuiComponentContext/MuiComponentContext';
+import { useMuiComponent } from '../MuiComponentContext';
 import AddButton from '../AddButton/AddButton';
 
 const { canExpand } = utils;
@@ -21,50 +21,39 @@ const ObjectFieldTemplate = ({
   formData,
   onAddClick,
 }: ObjectFieldTemplateProps) => {
-  const { Grid } = useContext(MuiComponentContext);
-  return <>
-    {(uiSchema['ui:title'] || title) && (
-      <TitleField
-        id={`${idSchema.$id}-title`}
-        title={title}
-        required={required}
-      />
-    )}
-    {description && (
-      <DescriptionField
-        id={`${idSchema.$id}-description`}
-        description={description}
-      />
-    )}
-    <Grid container={true} spacing={2} style={{ marginTop: '10px' }}>
-      {properties.map((element, index) =>
-        // Remove the <Grid> if the inner element is hidden as the <Grid>
-        // itself would otherwise still take up space.
-        element.hidden ? (
-          element.content
-        ) : (
-          <Grid
-            item={true}
-            xs={12}
-            key={index}
-            style={{ marginBottom: '10px' }}>
-            {element.content}
-          </Grid>
-        )
+  const { Grid } = useMuiComponent();
+  return (
+    <>
+      {(uiSchema['ui:title'] || title) && (
+        <TitleField id={`${idSchema.$id}-title`} title={title} required={required} />
       )}
-      {canExpand(schema, uiSchema, formData) && (
-        <Grid container justifyContent="flex-end">
-          <Grid item={true}>
-            <AddButton
-              className="object-property-expand"
-              onClick={onAddClick(schema)}
-              disabled={disabled || readonly}
-            />
+      {description && <DescriptionField id={`${idSchema.$id}-description`} description={description} />}
+      <Grid container={true} spacing={2} style={{ marginTop: '10px' }}>
+        {properties.map((element, index) =>
+          // Remove the <Grid> if the inner element is hidden as the <Grid>
+          // itself would otherwise still take up space.
+          element.hidden ? (
+            element.content
+          ) : (
+            <Grid item={true} xs={12} key={index} style={{ marginBottom: '10px' }}>
+              {element.content}
+            </Grid>
+          )
+        )}
+        {canExpand(schema, uiSchema, formData) && (
+          <Grid container justifyContent="flex-end">
+            <Grid item={true}>
+              <AddButton
+                className="object-property-expand"
+                onClick={onAddClick(schema)}
+                disabled={disabled || readonly}
+              />
+            </Grid>
           </Grid>
-        </Grid>
-      )}
-    </Grid>
-  </>;
+        )}
+      </Grid>
+    </>
+  );
 };
 
 export default ObjectFieldTemplate;