+++ packages/material-ui/test/Form.test.tsx
@@ -230,4 +230,11 @@ describe("single fields", () => {
       .toJSON();
     expect(tree).toMatchSnapshot();
   });
+  test("using custom tagName", () => {
+    const schema: JSONSchema7 = {
+      type: "string",
+    };
+    const tree = renderer.create(<Form schema={schema} tagName="div" />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });